requirements.txt
asgiref==3.7.2
astroid==2.15.6
black==23.7.0
blinker==1.6.2
click==8.1.6
dill==0.3.7
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
Flask==2.3.2
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomlkit==0.12.1
Werkzeug==2.3.6
wrapt==1.15.0
